+++ edx_proctoring/api.py
@@ -81,6 +81,14 @@ REJECTED_GRADE_OVERRIDE_EARNED = 0.0
 USER_MODEL = get_user_model()
 
 
+def get_total_allowed_time_for_exam(exam, user_id):
+    """
+    Returns total allowed time in humanized form for exam including allowance time.
+    """
+    total_allowed_time = _calculate_allowed_mins(exam, user_id)
+    return humanized_time(total_allowed_time)
+
+
 def get_proctoring_settings_by_exam_id(exam_id):
     """
     Return proctoring settings and exam proctoring backend for proctored exam by exam_id.
@@ -744,7 +752,6 @@ def get_exam_attempt_data(exam_id, attempt_id, is_learning_mfe=False):
     # a same process as the LMS
     if is_learning_mfe:
         exam_url_path = resolve_exam_url_for_learning_mfe(exam['course_id'], exam['content_id'])
-
     else:
         exam_url_path = reverse('jump_to', args=[exam['course_id'], exam['content_id']])